Issue #15 Put all of the .html pages into a DevUI folder.
This work also included a significant reworking of the two book*.html pages since these are the main Bible reader pages.  Those use a fairly small amount of HTML and a script that fetches the data and populates the html page.   The exact behavior is controlled by the BibleModel-preferences.json.

This allows the data to be fetched from a file stored in the defaultLibrary directory, or can be fetched from the REST server.

One major piece of work I will mention here is a lot of commenting in the bookWithContent.html page.   This describes the various sticky tricky bits of javascript.  It's quite a complex mess that I needed to understand and help other contributors.   Rather than duplicate much of this code in a number of pages, I plan to refactor it into a separate .mjs file that can be sourced into each page.  That refactoring will be another mini project.   Stay tuned for that.